스타일 컴포넌트 분리 + 주석으로 인한 함수 설명
