+++ src/components/MyPage/Button/Button.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import { StyledButton } from './style';
+
+interface Props {
+  children: React.ReactNode;
+  onClick: React.MouseEventHandler<HTMLButtonElement>;
+}
+
+const Button = ({ children, onClick }: Props) => {
+  return (
+    <div>
+      <StyledButton onClick={onClick}>{children}</StyledButton>
+    </div>
+  );
+};
+
+export default Button;
