+++ src/components/MyPage/ProfileTap/style.js
@@ -1,0 +1,20 @@
+import styled from 'styled-components';
+
+const StyledProfileDiv = styled.div`
+  @media screen and (max-width: 768px) {
+  }
+`;
+
+const StyledTextH4 = styled.h4`
+  color: #6a6969;
+  font-size: larger;
+  margin-bottom: 1rem;
+  text-align: center;
+`;
+
+const StyledTextP = styled.p`
+  color: #a7a7a7;
+  text-align: center;
+`;
+
+export { StyledProfileDiv, StyledTextH4, StyledTextP };
