+++ src/pages/MyPage/Mypage.tsx
@@ -2,11 +2,11 @@ import { getAuth } from 'firebase/auth';
 import { useEffect, useState } from 'react';
 import { useQuery } from 'react-query';
 import styled from 'styled-components';
-import Button from '../../components/MyPage/Button';
-import MyReview from '../../components/MyPage/MyReview';
-import Profile from '../../components/MyPage/Profile';
-import ProfileModal from '../../components/MyPage/ProfileModal';
-import ProfileTap from '../../components/MyPage/ProfileTap';
+import Button from '../../components/MyPage/Button/Button';
+import MyReview from '../../components/MyPage/myReview/MyReview';
+import Profile from '../../components/MyPage/Profile/Profile';
+import ProfileModal from '../../components/MyPage/ProfileModal/ProfileModal';
+import ProfileTap from '../../components/MyPage/ProfileTap/ProfileTap';
 
 export const Mypage = () => {
   // 모달창 노출 여부 state
